Rename python file with configuration tools
